remove now behaves correctly with dryrun/verbose
